Refactor "My Courses"
- Fix initial load issue mentioned in https://openedx.atlassian.net/browse/MA-2758 by moving the loadData(true) call to onViewCreated
- Add data binding
- Merge MyCourseListTabFragment with its abstract parent class, since it is the only implementation
