+++ VideoLocker/res/drawable/fg_course_row.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:drawable="@drawable/bg_frame_course" />
+    <item android:drawable="@drawable/selectable_rounded_box_overlay" />
+</layer-list>
