+++ VideoLocker/res/layout/fragment_my_all_videos.xml
@@ -1,37 +1,37 @@
 <?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android">
-<RelativeLayout
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/grey_act_background"
-    android:orientation="vertical"
-    android:splitMotionEvents="false" >
-    
-    <!-- keep this container, as playerfragment needs it to be present -->
-    <LinearLayout
-        android:id="@+id/videos_player_container"
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools">
+
+    <RelativeLayout
         android:layout_width="match_parent"
-        android:layout_height="200dp"
-        android:visibility="gone"
-        android:orientation="vertical" > 
-    </LinearLayout>
+        android:layout_height="match_parent"
+        android:background="@color/grey_act_background"
+        android:orientation="vertical"
+        android:splitMotionEvents="false">
 
-    <ListView
-        android:id="@+id/videos_course_list"
-        android:layout_below="@id/videos_player_container"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:divider="@android:color/transparent"
-        android:dividerHeight="10.0sp"
-        android:padding="10dp"
-        android:drawSelectorOnTop="false"
-		android:listSelector="@color/grey_act_background"
-        />
+        <!-- keep this container, as playerfragment needs it to be present -->
+        <LinearLayout
+            android:id="@+id/videos_player_container"
+            android:layout_width="match_parent"
+            android:layout_height="@dimen/video_player_height"
+            android:orientation="vertical"
+            android:visibility="gone" />
 
-    <TextView
-        android:id="@+id/empty_list_view"
-        android:text="@string/no_videos_to_display"
-        style="@style/content_unavailable_error_text" />
+        <ListView
+            android:id="@+id/videos_course_list"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_below="@id/videos_player_container"
+            android:divider="@android:color/transparent"
+            android:dividerHeight="@dimen/widget_margin"
+            android:drawSelectorOnTop="false"
+            android:listSelector="@android:color/transparent"
+            tools:listitem="@layout/row_myvideo_course_list" />
 
-</RelativeLayout>
+        <TextView
+            android:id="@+id/empty_list_view"
+            style="@style/content_unavailable_error_text"
+            android:text="@string/no_videos_to_display" />
+
+    </RelativeLayout>
 </layout>
