--- VideoLocker/res/layout/fragment_my_course_list_tab.xml
@@ -1,65 +1,0 @@
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <View
-        android:id="@+id/offline_bar"
-        style="@style/offline_bar"
-        android:layout_alignParentTop="true"
-        android:visibility="visible" />
-
-    <LinearLayout
-        android:id="@+id/no_course_tv"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:gravity="center"
-        android:orientation="vertical">
-
-        <org.edx.mobile.view.custom.IconImageViewXml
-            android:layout_width="90dp"
-            android:layout_height="90dp"
-            android:layout_margin="20dp"
-            app:iconColor="@color/edx_grayscale_neutral_light"
-            app:iconName="fa-exclamation-circle" />
-
-        <TextView
-            style="@style/content_unavailable_error_text"
-            android:text="@string/no_courses_to_display"
-            android:visibility="visible"/>
-    </LinearLayout>
-
-    <android.support.v4.widget.SwipeRefreshLayout
-        android:id="@+id/swipe_container"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_below="@id/offline_bar">
-
-        <ListView
-            tools:listitem="@layout/row_course_list"
-            android:id="@+id/my_course_list"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:divider="@null"
-            android:dividerHeight="0dp"
-            android:drawSelectorOnTop="false"
-            android:listSelector="@color/grey_act_background"
-            android:clipToPadding="false"/>
-    </android.support.v4.widget.SwipeRefreshLayout>
-
-    <include layout="@layout/loading_indicator" />
-
-    <LinearLayout
-        android:id="@+id/offline_panel"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:visibility="gone"
-        tools:visibility="visible">
-
-        <include layout="@layout/panel_offline_mode" />
-    </LinearLayout>
-
-</RelativeLayout>
-
-
